+++ src/main/java/com/warkoczewski/AgroAdventureBooking/model/entity/Role.java
@@ -1,6 +1,7 @@
-package com.warkoczewski.AgroAdventureBooking.model;
+package com.warkoczewski.AgroAdventureBooking.model.entity;
 
 
+import com.warkoczewski.AgroAdventureBooking.model.User;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
