+++ src/main/java/com/warkoczewski/AgroAdventureBooking/model/presentation/FarmPage.java
@@ -1,0 +1,4 @@
+package com.warkoczewski.AgroAdventureBooking.model.presentation;
+
+public class FarmPage {
+}
