+++ src/main/java/com/warkoczewski/AgroAdventureBooking/repository/BookingRepository.java
@@ -1,6 +1,6 @@
 package com.warkoczewski.AgroAdventureBooking.repository;
 
-import com.warkoczewski.AgroAdventureBooking.model.Booking;
+import com.warkoczewski.AgroAdventureBooking.model.entity.Booking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookingRepository extends JpaRepository<Booking, Long> {
