+++ src/main/java/com/warkoczewski/AgroAdventureBooking/repository/DateRepository.java
@@ -1,6 +1,6 @@
 package com.warkoczewski.AgroAdventureBooking.repository;
 
-import com.warkoczewski.AgroAdventureBooking.model.Date;
+import com.warkoczewski.AgroAdventureBooking.model.entity.Date;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDate;
