+++ src/main/java/com/warkoczewski/AgroAdventureBooking/repository/FarmRepository.java
@@ -1,14 +1,11 @@
 package com.warkoczewski.AgroAdventureBooking.repository;
 
-import com.warkoczewski.AgroAdventureBooking.model.Farm;
+import com.warkoczewski.AgroAdventureBooking.model.entity.Farm;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface FarmRepository extends JpaRepository<Farm, Long> {
     List<Farm> findAll();
