+++ src/main/java/com/warkoczewski/AgroAdventureBooking/service/BookingService.java
@@ -1,7 +1,7 @@
 package com.warkoczewski.AgroAdventureBooking.service;
 
 import com.warkoczewski.AgroAdventureBooking.dto.BookingDTO;
-import com.warkoczewski.AgroAdventureBooking.model.Booking;
+import com.warkoczewski.AgroAdventureBooking.model.entity.Booking;
 
 public interface BookingService {
     Booking createBooking(BookingDTO bookingDTO);
