+++ src/main/java/com/warkoczewski/AgroAdventureBooking/service/DateService.java
@@ -1,7 +1,7 @@
 package com.warkoczewski.AgroAdventureBooking.service;
 
 import com.warkoczewski.AgroAdventureBooking.dto.BookingDTO;
-import com.warkoczewski.AgroAdventureBooking.model.Date;
+import com.warkoczewski.AgroAdventureBooking.model.entity.Date;
 
 import java.util.List;
