+++ src/main/java/com/warkoczewski/AgroAdventureBooking/service/FarmService.java
@@ -4,7 +4,7 @@ import com.warkoczewski.AgroAdventureBooking.dto.AddressDTO;
 import com.warkoczewski.AgroAdventureBooking.dto.DisplayFarmDTO;
 import com.warkoczewski.AgroAdventureBooking.dto.FarmDTO;
 import com.warkoczewski.AgroAdventureBooking.dto.LocationDTO;
-import com.warkoczewski.AgroAdventureBooking.model.Farm;
+import com.warkoczewski.AgroAdventureBooking.model.entity.Farm;
 import org.springframework.data.domain.Page;
 
 import java.util.List;
