+++ src/main/java/com/warkoczewski/AgroAdventureBooking/service/impl/FarmServiceImpl.java
@@ -4,18 +4,15 @@ import com.warkoczewski.AgroAdventureBooking.dto.AddressDTO;
 import com.warkoczewski.AgroAdventureBooking.dto.DisplayFarmDTO;
 import com.warkoczewski.AgroAdventureBooking.dto.FarmDTO;
 import com.warkoczewski.AgroAdventureBooking.dto.LocationDTO;
-import com.warkoczewski.AgroAdventureBooking.model.Address;
-import com.warkoczewski.AgroAdventureBooking.model.Farm;
-import com.warkoczewski.AgroAdventureBooking.model.Location;
+import com.warkoczewski.AgroAdventureBooking.model.entity.Address;
+import com.warkoczewski.AgroAdventureBooking.model.entity.Farm;
+import com.warkoczewski.AgroAdventureBooking.model.entity.Location;
 import com.warkoczewski.AgroAdventureBooking.repository.FarmRepository;
 import com.warkoczewski.AgroAdventureBooking.service.FarmService;
-import org.modelmapper.Converter;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.jaxb.SpringDataJaxb;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
