+++ src/main/java/com/warkoczewski/AgroAdventureBooking/service/impl/LocationServiceImpl.java
@@ -2,7 +2,7 @@ package com.warkoczewski.AgroAdventureBooking.service.impl;
 
 import com.warkoczewski.AgroAdventureBooking.dto.LocationDTO;
 import com.warkoczewski.AgroAdventureBooking.exception.DuplicateCoordinatesException;
-import com.warkoczewski.AgroAdventureBooking.model.Location;
+import com.warkoczewski.AgroAdventureBooking.model.entity.Location;
 import com.warkoczewski.AgroAdventureBooking.repository.LocationRepository;
 import com.warkoczewski.AgroAdventureBooking.service.LocationService;
 import lombok.extern.slf4j.Slf4j;
